- upgrade wordpress from 4.9.25 to 6.4.3 - upgrade plugin SyntaxHighlighter Evolved plugin from 3.2.1 to 3.7.0 - upgrade plugin Yoast SEO from 7.2 to 22.1

+++ wp-content/plugins/wordpress-seo/admin/import/plugins/class-abstract-plugin-importer.php
@@ -6,11 +6,12 @@
  */
 
 /**
- * Class WPSEO_Plugin_Importer
+ * Class WPSEO_Plugin_Importer.
  *
  * Class with functionality to import meta data from other plugins.
  */
 abstract class WPSEO_Plugin_Importer {
+
 	/**
 	 * Holds the import status object.
 	 *
@@ -26,17 +27,23 @@ abstract class WPSEO_Plugin_Importer {
 	protected $plugin_name;
 
 	/**
-	 * Meta key, used in SQL LIKE clause for detect query.
+	 * Meta key, used in SQL LIKE clause for delete query.
 	 *
 	 * @var string
 	 */
 	protected $meta_key;
 
 	/**
+	 * Array of meta keys to detect and import.
+	 *
+	 * @var array
+	 */
+	protected $clone_keys;
+
+	/**
 	 * Class constructor.
 	 */
-	public function __construct() {
-	}
+	public function __construct() {}
 
 	/**
 	 * Returns the string for the plugin we're importing from.
@@ -98,6 +105,9 @@ abstract class WPSEO_Plugin_Importer {
 	 */
 	protected function cleanup() {
 		global $wpdb;
+		if ( empty( $this->meta_key ) ) {
+			return true;
+		}
 		$wpdb->query(
 			$wpdb->prepare(
 				"DELETE FROM {$wpdb->postmeta} WHERE meta_key LIKE %s",
@@ -106,11 +116,20 @@ abstract class WPSEO_Plugin_Importer {
 		);
 		$result = $wpdb->__get( 'result' );
 		if ( ! $result ) {
-			/* translators: %s is replaced with the plugin's name. */
-			$this->status->set_msg( sprintf( __( 'Cleanup of %s data failed.', 'wordpress-seo' ), $this->plugin_name ) );
+			$this->cleanup_error_msg();
 		}
 
 		return $result;
+	}
+
+	/**
+	 * Sets the status message for when a cleanup has gone bad.
+	 *
+	 * @return void
+	 */
+	protected function cleanup_error_msg() {
+		/* translators: %s is replaced with the plugin's name. */
+		$this->status->set_msg( sprintf( __( 'Cleanup of %s data failed.', 'wordpress-seo' ), $this->plugin_name ) );
 	}
 
 	/**
@@ -135,12 +154,17 @@ abstract class WPSEO_Plugin_Importer {
 	 */
 	protected function detect() {
 		global $wpdb;
-		$result = $wpdb->get_var(
+
+		$meta_keys = wp_list_pluck( $this->clone_keys, 'old_key' );
+		$result    = $wpdb->get_var(
 			$wpdb->prepare(
-				"SELECT COUNT(*) AS `count` FROM {$wpdb->postmeta} WHERE meta_key LIKE %s",
-				$this->meta_key
+				"SELECT COUNT(*) AS `count`
+					FROM {$wpdb->postmeta}
+					WHERE meta_key IN ( " . implode( ', ', array_fill( 0, count( $meta_keys ), '%s' ) ) . ' )',
+				$meta_keys
 			)
 		);
+
 		if ( $result === '0' ) {
 			return false;
 		}
@@ -157,12 +181,13 @@ abstract class WPSEO_Plugin_Importer {
 	 *
 	 * @return bool Clone status.
 	 */
-	protected function meta_key_clone( $old_key, $new_key, $replace_values = array() ) {
+	protected function meta_key_clone( $old_key, $new_key, $replace_values = [] ) {
 		global $wpdb;
 
 		// First we create a temp table with all the values for meta_key.
 		$result = $wpdb->query(
 			$wpdb->prepare(
+				// phpcs:ignore WordPress.DB.DirectDatabaseQuery.SchemaChange -- This is intentional + temporary.
 				"CREATE TEMPORARY TABLE tmp_meta_table SELECT * FROM {$wpdb->postmeta} WHERE meta_key = %s",
 				$old_key
 			)
@@ -171,6 +196,7 @@ abstract class WPSEO_Plugin_Importer {
 			$this->set_missing_db_rights_status();
 			return false;
 		}
+
 		// Delete all the values in our temp table for posts that already have data for $new_key.
 		$wpdb->query(
 			$wpdb->prepare(
@@ -179,7 +205,10 @@ abstract class WPSEO_Plugin_Importer {
 			)
 		);
 
-		// We set meta_id to NULL so on re-insert into the postmeta table, MYSQL can set new meta_id's and we don't get duplicates.
+		/*
+		 * We set meta_id to NULL so on re-insert into the postmeta table, MYSQL can set
+		 * new meta_id's and we don't get duplicates.
+		 */
 		$wpdb->query( 'UPDATE tmp_meta_table SET meta_id = NULL' );
 
 		// Now we rename the meta_key.
@@ -196,6 +225,7 @@ abstract class WPSEO_Plugin_Importer {
 		$wpdb->query( "INSERT INTO {$wpdb->postmeta} SELECT * FROM tmp_meta_table" );
 
 		// Now we drop our temporary table.
+		// phpcs:ignore WordPress.DB.DirectDatabaseQuery.SchemaChange -- This is intentional + a temporary table.
 		$wpdb->query( 'DROP TEMPORARY TABLE IF EXISTS tmp_meta_table' );
 
 		return true;
@@ -210,7 +240,7 @@ abstract class WPSEO_Plugin_Importer {
 	 */
 	protected function meta_keys_clone( $clone_keys ) {
 		foreach ( $clone_keys as $clone_key ) {
-			$result = $this->meta_key_clone( $clone_key['old_key'], $clone_key['new_key'], isset( $clone_key['convert'] ) ? $clone_key['convert'] : array() );
+			$result = $this->meta_key_clone( $clone_key['old_key'], $clone_key['new_key'], ( $clone_key['convert'] ?? [] ) );
 			if ( ! $result ) {
 				return false;
 			}
@@ -220,6 +250,8 @@ abstract class WPSEO_Plugin_Importer {
 
 	/**
 	 * Sets the import status to false and returns a message about why it failed.
+	 *
+	 * @return void
 	 */
 	protected function set_missing_db_rights_status() {
 		$this->status->set_status( false );
@@ -228,16 +260,46 @@ abstract class WPSEO_Plugin_Importer {
 	}
 
 	/**
+	 * Helper function to search for a key in an array and maybe save it as a meta field.
+	 *
+	 * @param string $plugin_key The key in the $data array to check.
+	 * @param string $yoast_key  The identifier we use in our meta settings.
+	 * @param array  $data       The array of data for this post to sift through.
+	 * @param int    $post_id    The post ID.
+	 *
+	 * @return void
+	 */
+	protected function import_meta_helper( $plugin_key, $yoast_key, $data, $post_id ) {
+		if ( ! empty( $data[ $plugin_key ] ) ) {
+			$this->maybe_save_post_meta( $yoast_key, $data[ $plugin_key ], $post_id );
+		}
+	}
+
+	/**
 	 * Saves a post meta value if it doesn't already exist.
 	 *
 	 * @param string $new_key The key to save.
 	 * @param mixed  $value   The value to set the key to.
 	 * @param int    $post_id The Post to save the meta for.
+	 *
+	 * @return void
 	 */
 	protected function maybe_save_post_meta( $new_key, $value, $post_id ) {
-		$existing_value = get_post_meta( $post_id, WPSEO_Meta::$meta_prefix . $new_key, true );
+		// Big. Fat. Sigh. Mostly used for _yst_is_cornerstone, but might be useful for other hidden meta's.
+		$key        = WPSEO_Meta::$meta_prefix . $new_key;
+		$wpseo_meta = true;
+		if ( substr( $new_key, 0, 1 ) === '_' ) {
+			$key        = $new_key;
+			$wpseo_meta = false;
+		}
+
+		$existing_value = get_post_meta( $post_id, $key, true );
 		if ( empty( $existing_value ) ) {
-			WPSEO_Meta::set_value( $new_key, $value, $post_id );
+			if ( $wpseo_meta ) {
+				WPSEO_Meta::set_value( $new_key, $value, $post_id );
+				return;
+			}
+			update_post_meta( $post_id, $new_key, $value );
 		}
 	}
 
@@ -252,7 +314,7 @@ abstract class WPSEO_Plugin_Importer {
 		global $wpdb;
 
 		// Now we replace values if needed.
-		if ( is_array( $replace_values ) && $replace_values !== array() ) {
+		if ( is_array( $replace_values ) && $replace_values !== [] ) {
 			foreach ( $replace_values as $old_value => $new_value ) {
 				$wpdb->query(
 					$wpdb->prepare(
